fix: killswitch and frozen flag

+++ rlp/programs/rlp/src/states/killswitch.rs
@@ -6,43 +6,43 @@ use crate::errors::RlpError;
 #[repr(C)]
 #[derive(BorshSchema, AnchorSerialize, AnchorDeserialize, Clone, Copy, Debug, PartialEq, Eq, InitSpace, Default)]
 pub struct KillSwitch {
-    pub frozen: u8,
+    pub frozen: u16,
 }
 
 impl KillSwitch {
     pub fn deserialize(buf: &mut &[u8]) -> Result<Self> {
-        let frozen = u8::deserialize(buf)?;
+        let frozen = u16::deserialize(buf)?;
         Ok(KillSwitch { frozen })
     }
-    
+
     pub fn try_serialise<W: Write>(&self, writer: &mut W) -> Result<()> {
         self.frozen.serialize(writer)?;
         Ok(())
     }
-    
+
     /** Check if an action is frozen. */
     pub fn is_frozen(&self, action: &Action) -> bool {
-        let mask = 1u8 << (*action as u8);
+        let mask = 1u16 << (*action as u16);
         (self.frozen & mask) != 0
-    }    
+    }
 
     /** Throw if action can't be performed. */
     pub fn action_unsuspended(&self, action: &Action) -> Result<()> {
-        if !self.is_frozen(action) { 
-            return Ok(()); 
-        }        
+        if !self.is_frozen(action) {
+            return Ok(());
+        }
         Err(error!(RlpError::ActionFrozen))
     }
 
     /** Freeze an action. */
     pub fn freeze(&mut self, action: &Action) {
-        let mask = 1u8 << (*action as u8);
+        let mask = 1u16 << (*action as u16);
         self.frozen |= mask;
     }
 
     /** Unfreeze an action. */
     pub fn unfreeze(&mut self, action: &Action) {
-        let mask = 1u8 << (*action as u8);
+        let mask = 1u16 << (*action as u16);
         self.frozen &= !mask;
     }
 }
